requirements.txt
streamlit
pandas
openpyxl
python-dotenv
openai
numpy
scikit-learn